CourseScrape/requirements.txt
aiohappyeyeballs     
aiohttp              
aiosignal            
asgiref              
attrs                
beautifulsoup4       
bs4                  
certifi              
charset-normalizer   
click                
colorama             
Django               
django-rest-framework
djangorestframework  
dnspython            
dotenv               
duckduckgo_search    
filelock             
frozenlist           
fsspec               
greenlet             
huggingface-hub      
idna                 
Jinja2               
lxml                 
Markdown             
MarkupSafe           
mpmath               
multidict            
nest-asyncio         
networkx             
numpy                
packaging            
pip                  
playwright           
primp                
propcache            
psycopg2-binary      
pyee                 
pymongo              
python-dotenv        
PyYAML               
regex                
requests             
safetensors          
setuptools           
soupsieve            
sqlparse             
sympy                
tokenizers           
torch                
tqdm                 
transformers         
typing_extensions    
tzdata               
urllib3              
yarl       
huggingface-hub>=0.19.0          